requirements.txt
flake8==6.0.0
grpcio==1.51.1
grpcio-tools==1.51.1
mccabe==0.7.0
protobuf==4.21.12
pycodestyle==2.10.0
pyflakes==3.0.1
